seguimiento de clientes terminado, diseño de interfaz grafica de clientes, pagos y seguimiento terminado. uso de datatable en seguimiento implementado. cargado nuevos estilos css

+++ database/migrations/2022_02_14_023751_create_clientes_table.php
@@ -15,7 +15,7 @@ class CreateClientesTable extends Migration
     {
         Schema::create('clientes', function (Blueprint $table) {
             $table->id();
-            $table->integer('ide')->unique();
+            $table->bigInteger('ide')->unique();
             $table->string('nombre',40);
             $table->string('apellido',40);
             $table->string('telefono');
